Move UI factory concerns into top-level view controller.
